refactor(stream): align format option boundaries
Why:
- Review found that stream entry format option parsing belonged with the rest of CLI option normalization rather than the streaming writer boundary.

What Changed:
- Move stream entry format codec and option normalization types into Hosting/Cli/Options.
- Keep streaming writer and sink types in Common/Streaming as output infrastructure.
- Make NullCommandProgressSink ignore entry data while still honoring cancellation.

Impact:
- --format parsing follows the existing option-normalizer directory boundary, and no-op progress sinks no longer turn ignored entry data into execution failures.
